Change the ordering of pre-release identifiers
This changes the ordering of alphanumeric identifiers in versions so
that some identifiers are not ordered lexicographically. Identifiers
that are handled differently are: (SNAP, SNAPSHOT) < ALPHA < BETA < M < RC.
The motivation for this change is to report a version like 3.1.0-RC3
as an update for 3.1.0-SNAP13.

The change from `Int` to `BigInt` in `parsePart` has been made so that
the pre-release identifier in 1.0.0-20131213005945 is treated as numeric.
It is currently not because `"20131213005945".toInt` throws an exception.
Without this change, the version above would be greater than 1.0.0-alpha
which would lead to a test failure.

Closes #128.